Root user no deletion feature in UI and Makefile for lenghty commands

+++ app/blueprints/main.py
@@ -2,6 +2,7 @@
 from flask import request,redirect, url_for, Blueprint, render_template, flash  
 from app.models import User
 from app import db
+import re
 from datetime import datetime
 
 main = Blueprint('main', __name__)
@@ -15,7 +16,7 @@ def index():
 @main.route('/ui')
 def ui():
     
-    users = User.query.all()
+    users = User.query.order_by(User.username.asc()).all()
     return render_template('ui.html', title="UI Page", message="Welcome to the UI Page!", users=users)
 
 #CREATE
@@ -25,18 +26,33 @@ def create_user_ui():
     username=request.form['username']
     email=request.form['email']
     password=request.form['password']
-    new_user = User(username=username, email=email, password=generate_password_hash(password))
-    db.session.add(new_user)
-    db.session.commit()
-    flash('User created successfully!', 'success')
+    
+    if not re.match(r"[^@]+@[^@]+\.[^@]+", email):
+        flash('Invalid email address.', 'danger')
+        return redirect(url_for('main.ui'))
+    try:
+        new_user = User(username=username, email=email, password=password)
+        db.session.add(new_user)
+        db.session.commit()
+        flash('User created successfully!', 'success')
+    except Exception as e:
+        db.session.rollback()
+        flash(f'Error creating user: {str(e)}', 'danger')
     return redirect(url_for('main.ui'))
 
 #EDIT
 @main.route('/ui/edit_user/<int:user_id>', methods=['POST'])
 def edit_user_ui(user_id):
     user = User.query.get_or_404(user_id)
-    user.username = request.form['username']
-    user.email = request.form['email']
+    
+    new_email = request.form['email']
+    new_password = request.form.get('password') # Optional password update
+    
+    if new_email:
+        user.email = new_email
+    if new_password:
+        user.set_password(new_password)
+
     db.session.commit()
     flash('User updated successfully!', 'success')
     return redirect(url_for('main.ui'))
@@ -45,6 +61,11 @@ def edit_user_ui(user_id):
 @main.route('/ui/delete_user/<int:user_id>', methods=['POST'])
 def delete_user_ui(user_id):
     user = User.query.get_or_404(user_id)
+    
+    if user.is_root:
+        flash('Cannot delete root user!', 'danger')
+        return redirect(url_for('main.ui'))
+    
     db.session.delete(user)
     db.session.commit()
     flash('User deleted successfully!', 'success')
